+++ src/components/ui/legal-theme-types.ts
@@ -1,0 +1,12 @@
+export interface LegalThemeConfig {
+  enableAnimations: boolean
+  enableRippleEffect: boolean
+  contrastMode: 'normal' | 'high'
+  reducedMotion: boolean
+}
+
+export interface LegalThemeContextType {
+  config: LegalThemeConfig
+  updateConfig: (newConfig: Partial<LegalThemeConfig>) => void
+}
+
